# IAE.Web/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "IAE.Web, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Avaliacao/{id}": {
      "get": {
        "tags": [
          "Avaliacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/Avaliacao/Turma/{id}": {
      "get": {
        "tags": [
          "Avaliacao"
        ],
        "parameters": [
          {
            "name": "idTurma",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/Avaliacao/Professor/{id}": {
      "get": {
        "tags": [
          "Avaliacao"
        ],
        "parameters": [
          {
            "name": "idProfessor",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/Avaliacao/GerarSimulado": {
      "post": {
        "tags": [
          "Avaliacao"
        ],
        "summary": "Obter um Simulado",
        "parameters": [
          {
            "name": "turmaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "numeroQuestoes",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/Avaliacao/GerarProva": {
      "post": {
        "tags": [
          "Avaliacao"
        ],
        "summary": "Obter uma Prova",
        "parameters": [
          {
            "name": "turmaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "numeroQuestoes",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Avaliacao"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/PlanoEnsino/{id}": {
      "get": {
        "tags": [
          "PlanoEnsino"
        ],
        "parameters": [
          {
            "name": "idPlanoEnsino",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PlanoEnsino"
        ],
        "parameters": [
          {
            "name": "idPlanoEnsino",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoEnsinoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoEnsinoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoEnsinoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PlanoEnsino"
        ],
        "parameters": [
          {
            "name": "idPlanoEnsino",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/PlanoEnsino": {
      "get": {
        "tags": [
          "PlanoEnsino"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanoEnsino"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanoEnsino"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanoEnsino"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PlanoEnsino"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoEnsinoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoEnsinoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlanoEnsinoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanoEnsino"
                }
              }
            }
          }
        }
      }
    },
    "/api/PlanoEnsino/ObtePorIds": {
      "post": {
        "tags": [
          "PlanoEnsino"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanoEnsino"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanoEnsino"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanoEnsino"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Turma/{id}": {
      "get": {
        "tags": [
          "Turma"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Turma"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Turma"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Turma"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Turma"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Turma"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Turma"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Turma"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Turma"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Avaliacao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tipoAvaliacao": {
            "$ref": "#/components/schemas/TipoAvaliacao"
          },
          "idTurma": {
            "type": "integer",
            "format": "int32"
          },
          "idProfessor": {
            "type": "integer",
            "format": "int32"
          },
          "idsQuestoes": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "turma": {
            "$ref": "#/components/schemas/Turma"
          },
          "professor": {
            "$ref": "#/components/schemas/Usuario"
          },
          "questoes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Questao"
            },
            "nullable": true
          },
          "alunosParticipantes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Usuario"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Nota": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idProva": {
            "type": "integer",
            "format": "int32"
          },
          "valorNota": {
            "type": "number",
            "format": "double"
          },
          "idAluno": {
            "type": "integer",
            "format": "int32"
          },
          "tipoAvaliacao": {
            "$ref": "#/components/schemas/TipoAvaliacao"
          },
          "idProfessor": {
            "type": "integer",
            "format": "int32"
          },
          "idTurma": {
            "type": "integer",
            "format": "int32"
          },
          "prova": {
            "$ref": "#/components/schemas/Avaliacao"
          },
          "aluno": {
            "$ref": "#/components/schemas/Usuario"
          },
          "turma": {
            "$ref": "#/components/schemas/Turma"
          }
        },
        "additionalProperties": false
      },
      "PlanoEnsino": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "nomeDisciplina": {
            "type": "string",
            "nullable": true
          },
          "escolaridade": {
            "type": "string",
            "nullable": true
          },
          "cargaHoraria": {
            "type": "number",
            "format": "double"
          },
          "referenciasBibliograficas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReferenciaBibliografica"
            },
            "nullable": true
          },
          "topicosAbordados": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlanoEnsinoDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "nomeDisciplina": {
            "type": "string",
            "nullable": true
          },
          "escolaridade": {
            "type": "string",
            "nullable": true
          },
          "cargaHoraria": {
            "type": "number",
            "format": "double"
          },
          "referenciasBibliograficas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReferenciaBibliografica"
            },
            "nullable": true
          },
          "topicosAbordados": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Questao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "enunciado": {
            "type": "string",
            "nullable": true
          },
          "respostas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resposta"
            },
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReferenciaBibliografica": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "referencia": {
            "type": "string",
            "nullable": true
          },
          "planosEnsino": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanoEnsino"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Resposta": {
        "type": "object",
        "properties": {
          "enunciadoResposta": {
            "type": "string",
            "nullable": true
          },
          "isRespostaCorreta": {
            "type": "boolean"
          },
          "idQuestao": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TipoAvaliacao": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "TipoUsuario": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Turma": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "codigoTurma": {
            "type": "string",
            "nullable": true
          },
          "idProfessor": {
            "type": "integer",
            "format": "int32"
          },
          "idPlanoEnsino": {
            "type": "integer",
            "format": "int32"
          },
          "professor": {
            "$ref": "#/components/schemas/Usuario"
          },
          "alunos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Usuario"
            },
            "nullable": true
          },
          "planoEnsino": {
            "$ref": "#/components/schemas/PlanoEnsino"
          }
        },
        "additionalProperties": false
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "sobrenome": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          },
          "tipoUsuario": {
            "$ref": "#/components/schemas/TipoUsuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "turmas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Turma"
            },
            "nullable": true
          },
          "notas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Nota"
            },
            "nullable": true
          },
          "planosEnsino": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanoEnsino"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}